Move activeCell logic to selection store
